Updates
added case for when a user does not input the valid menu options for 1 and 2 to select searching for high or lower than weight. also got rid of the data member Parcel* next in the parcel struct because it was not used in the code. current adding the header comments